move zarr tests
